Reorganizing tests
It was needed because of changes in code
